requirements.txt
# Core Web Framework
fastapi>=0.115.0
uvicorn>=0.38.0
starlette>=0.40.0
pydantic>=2.11.0
pydantic-settings>=2.10.0
python-multipart>=0.0.20
sse-starlette>=3.0.0

# LangChain & LLM Integration
langchain>=0.3.27
langchain-core>=0.3.75
langchain-community>=0.3.29
langchain-mistralai>=0.2.11
langgraph>=0.6.6
langgraph-checkpoint>=3.0.1
langgraph-prebuilt>=0.6.4
langgraph-checkpoint-sqlite>=3.0.0
mistralai>=1.9.0

# Tracing & Observability
langsmith>=0.4.0

# Database
aiosqlite>=0.21.0
SQLAlchemy>=2.0.0

# Environment & Configuration
python-dotenv>=1.2.0

# HTTP Clients
aiohttp>=3.12.0
requests>=2.32.0
httpx>=0.28.0

# Utilities
tenacity>=9.1.0
typing-extensions>=4.15.0
anyio>=4.7.0
attrs>=25.0.0
certifi>=2025.0.0
charset-normalizer>=3.4.0
idna>=3.10
urllib3>=2.5.0
frozenlist>=1.7.0
aiosignal>=1.4.0
multidict>=6.6.0
yarl>=1.20.0
